requirements.txt
python-dotenv==1.0.0
Markdown==3.4.4

# Django dependencies
django==4.2.4
djangorestframework==3.14.0
django-cors-headers==4.2.0
requests==2.31.0

# Integrations
medsenger-api==0.1.59

# Production deploy
psycopg2-binary==2.9.7
sentry-sdk[django]